+++ core/src/main/java/org/apache/metamodel/membrane/controllers/TenantController.java
@@ -18,16 +18,17 @@
  */
 package org.apache.metamodel.membrane.controllers;
 
-import java.util.LinkedHashMap;
+import java.util.Collections;
 import java.util.List;
-import java.util.Map;
 import java.util.stream.Collectors;
 
 import javax.ws.rs.core.UriBuilder;
 
 import org.apache.metamodel.membrane.app.TenantContext;
 import org.apache.metamodel.membrane.app.TenantRegistry;
-import org.apache.metamodel.membrane.controllers.model.RestLink;
+import org.apache.metamodel.membrane.swagger.model.DeleteTenantResponse;
+import org.apache.metamodel.membrane.swagger.model.GetTenantResponse;
+import org.apache.metamodel.membrane.swagger.model.GetTenantResponseDatasources;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -49,46 +50,48 @@ public class TenantController {
 
     @RequestMapping(method = RequestMethod.GET)
     @ResponseBody
-    public Map<String, Object> getTenant(@PathVariable("tenant") String tenantName) {
+    public GetTenantResponse getTenant(@PathVariable("tenant") String tenantName) {
         final TenantContext tenantContext = tenantRegistry.getTenantContext(tenantName);
         final String tenantNameNormalized = tenantContext.getTenantName();
 
         final UriBuilder uriBuilder = UriBuilder.fromPath("/{tenant}/{datasource}");
 
         final List<String> dataContextIdentifiers = tenantContext.getDataSourceRegistry().getDataSourceNames();
-        final List<RestLink> dataSourceLinks = dataContextIdentifiers.stream().map(s -> new RestLink(s, uriBuilder
-                .build(tenantNameNormalized, s))).collect(Collectors.toList());
+        final List<GetTenantResponseDatasources> dataSourceLinks = dataContextIdentifiers.stream().map(s -> {
+            final String uri = uriBuilder.build(tenantNameNormalized, s).toString();
+            return new GetTenantResponseDatasources().name(s).uri(uri);
+        }).collect(Collectors.toList());
 
-        final Map<String, Object> map = new LinkedHashMap<>();
-        map.put("type", "tenant");
-        map.put("name", tenantNameNormalized);
-        map.put("datasources", dataSourceLinks);
-        return map;
+        final GetTenantResponse resp = new GetTenantResponse();
+        resp.type("tenant");
+        resp.name(tenantNameNormalized);
+        resp.datasources(dataSourceLinks);
+        return resp;
     }
 
     @RequestMapping(method = RequestMethod.PUT)
     @ResponseBody
-    public Map<String, Object> putTenant(@PathVariable("tenant") String tenantName) {
+    public GetTenantResponse putTenant(@PathVariable("tenant") String tenantName) {
         final TenantContext tenantContext = tenantRegistry.createTenantContext(tenantName);
         final String tenantIdentifier = tenantContext.getTenantName();
 
-        final Map<String, Object> map = new LinkedHashMap<>();
-        map.put("type", "tenant");
-        map.put("name", tenantIdentifier);
-
-        return map;
+        final GetTenantResponse resp = new GetTenantResponse();
+        resp.type("tenant");
+        resp.name(tenantIdentifier);
+        resp.datasources(Collections.emptyList());
+        return resp;
     }
 
     @RequestMapping(method = RequestMethod.DELETE)
     @ResponseBody
-    public Map<String, Object> deleteTenant(@PathVariable("tenant") String tenantName) {
+    public DeleteTenantResponse deleteTenant(@PathVariable("tenant") String tenantName) {
         tenantRegistry.deleteTenantContext(tenantName);
 
-        final Map<String, Object> map = new LinkedHashMap<>();
-        map.put("type", "tenant");
-        map.put("name", tenantName);
-        map.put("deleted", true);
+        final DeleteTenantResponse resp = new DeleteTenantResponse();
+        resp.type("tenant");
+        resp.name(tenantName);
+        resp.deleted(true);
 
-        return map;
+        return resp;
     }
 }