+++ CHANGES.md
@@ -1,3 +1,4 @@
 ### Apache MetaModel Membrane 0.1 [work in progress]
 
+* [METAMODEL-1147] - Implemented Swagger codegen to build model classes from swagger file.
 * Established project source control, structure and build.